+++ CHANGELOG.md
@@ -6,4 +6,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.0.1]
 - Initial release
+- Supports endpoint testing
+
+## [0.0.2]
+### Added
+- Added an new command "APIi: New Inspection from WSDL"
+- Added support for creating inspections from WSDLs